+++ src/main/java/kr/hhplus/be/server/ServerApplication.java
@@ -1,6 +1,6 @@
 package kr.hhplus.be.server;
 
-import kr.hhplus.be.server.common.security.JwtProperties;
+import kr.hhplus.be.server.common.security.jwt.JwtProperties;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
